Made cortex-processing a command-line tool

+++ cortex-processing/src/main/kotlin/com/hileco/cortex/processing/geth/GethContractLoader.kt
@@ -4,7 +4,7 @@ import com.hileco.cortex.processing.database.NetworkModel
 import java.math.BigDecimal
 
 class GethContractLoader : GethLoader() {
-    fun load(networkModel: NetworkModel, blockNumberStart: BigDecimal, blockNumberEnd: BigDecimal): List<GethContract> {
+    fun load(networkModel: NetworkModel, blockNumberStart: Int, blockNumberEnd: Int): List<GethContract> {
         val result = executeGeth("load-contracts.js", networkModel.networkAddress, blockNumberStart, blockNumberEnd)
         return result.elements()
                 .asSequence()
